// src/graphql/videoCall/types/LeaveCall.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: LeaveCall
// ====================================================

export interface LeaveCall {
  leaveCall: string;
}

export interface LeaveCallVariables {
  token: string;
  userId: number;
}